+++ src/main/java/me/melontini/commander/impl/expression/functions/StructContainsKeyFunction.java
@@ -1,19 +1,23 @@
 package me.melontini.commander.impl.expression.functions;
 
+import com.ezylang.evalex.EvaluationContext;
 import com.ezylang.evalex.EvaluationException;
 import com.ezylang.evalex.Expression;
 import com.ezylang.evalex.data.EvaluationValue;
 import com.ezylang.evalex.functions.AbstractFunction;
 import com.ezylang.evalex.functions.FunctionParameter;
 import com.ezylang.evalex.parser.ASTNode;
-import com.ezylang.evalex.parser.ParseException;
 import com.ezylang.evalex.parser.Token;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.List;
 
 @FunctionParameter(name = "struct")
 @FunctionParameter(name = "key", isVarArg = true)
-public class StructContainsKeyFunction extends AbstractFunction implements CustomInlinerFunction {
+public class StructContainsKeyFunction extends AbstractFunction {
+
     @Override
-    public EvaluationValue evaluate(Expression expression, Token functionToken, EvaluationValue... par) throws EvaluationException {
+    public EvaluationValue evaluate(EvaluationContext context, Token functionToken, EvaluationValue... par) throws EvaluationException {
         return switch (par.length) {
             case 1 -> EvaluationValue.TRUE;
             case 2 -> par[0].getStructureValue().containsKey(par[1].getStringValue()) ? EvaluationValue.TRUE : EvaluationValue.FALSE;
@@ -29,7 +33,7 @@ public class StructContainsKeyFunction extends AbstractFunction implements Custo
     }
 
     @Override
-    public EvaluationValue cmd$inlineFunction(Expression expression, ASTNode node) throws ParseException, EvaluationException {
+    public @Nullable EvaluationValue inlineFunction(Expression expression, Token token, List<ASTNode> parameters) throws EvaluationException {
         return null;
     }
 }