+++ src/main/java/me/melontini/commander/impl/expression/functions/arrays/ArrayFind.java
@@ -1,14 +1,14 @@
 package me.melontini.commander.impl.expression.functions.arrays;
 
+import com.ezylang.evalex.EvaluationContext;
 import com.ezylang.evalex.EvaluationException;
 import com.ezylang.evalex.Expression;
 import com.ezylang.evalex.data.EvaluationValue;
 import com.ezylang.evalex.functions.AbstractFunction;
 import com.ezylang.evalex.functions.FunctionParameter;
 import com.ezylang.evalex.parser.ASTNode;
-import com.ezylang.evalex.parser.ParseException;
 import com.ezylang.evalex.parser.Token;
-import me.melontini.commander.impl.expression.functions.CustomInlinerFunction;
+import me.melontini.dark_matter.api.base.util.Exceptions;
 
 import java.util.List;
 
@@ -16,20 +16,18 @@ import static me.melontini.commander.impl.expression.EvalUtils.runLambda;
 
 @FunctionParameter(name = "array")
 @FunctionParameter(name = "predicate", isLazy = true)
-public class ArrayFind extends AbstractFunction implements CustomInlinerFunction {
+public class ArrayFind extends AbstractFunction {
 
     @Override
-    public EvaluationValue evaluate(Expression expression, Token functionToken, EvaluationValue... par) throws EvaluationException {
+    public EvaluationValue evaluate(EvaluationContext context, Token functionToken, EvaluationValue... par) throws EvaluationException {
         List<EvaluationValue> array = par[0].getArrayValue();
         ASTNode predicate = par[1].getExpressionNode();
 
-        return EvaluationValue.arrayValue(array.stream().filter(value -> runLambda(expression, value, predicate).getBooleanValue()).toList());
+        return EvaluationValue.arrayValue(array.stream().filter(value -> Exceptions.supply(() -> runLambda(context, value, predicate)).getBooleanValue()).toList());
     }
 
     @Override
-    public EvaluationValue cmd$inlineFunction(Expression expression, ASTNode node) throws ParseException, EvaluationException {
-        var value = CustomInlinerFunction.getNodeValue(node.getParameters().get(0));
-        if (value == null) return null;
-        return evaluate(expression, node.getToken(), value, EvaluationValue.expressionNodeValue(node.getParameters().get(1)));
+    public EvaluationValue inlineFunction(Expression expression, Token token, List<ASTNode> parameters) throws EvaluationException {
+        return null;
     }
 }