+++ src/main/java/me/melontini/commander/impl/expression/functions/arrays/ArrayMap.java
@@ -1,15 +1,15 @@
 package me.melontini.commander.impl.expression.functions.arrays;
 
+import com.ezylang.evalex.EvaluationContext;
 import com.ezylang.evalex.EvaluationException;
 import com.ezylang.evalex.Expression;
 import com.ezylang.evalex.data.EvaluationValue;
 import com.ezylang.evalex.functions.AbstractFunction;
 import com.ezylang.evalex.functions.FunctionParameter;
 import com.ezylang.evalex.parser.ASTNode;
-import com.ezylang.evalex.parser.ParseException;
 import com.ezylang.evalex.parser.Token;
 import com.google.common.collect.Lists;
-import me.melontini.commander.impl.expression.functions.CustomInlinerFunction;
+import me.melontini.dark_matter.api.base.util.Exceptions;
 
 import java.util.List;
 
@@ -17,20 +17,18 @@ import static me.melontini.commander.impl.expression.EvalUtils.runLambda;
 
 @FunctionParameter(name = "array")
 @FunctionParameter(name = "function", isLazy = true)
-public class ArrayMap extends AbstractFunction implements CustomInlinerFunction {
+public class ArrayMap extends AbstractFunction {
 
     @Override
-    public EvaluationValue evaluate(Expression expression, Token functionToken, EvaluationValue... par) throws EvaluationException {
+    public EvaluationValue evaluate(EvaluationContext expression, Token functionToken, EvaluationValue... par) throws EvaluationException {
         List<EvaluationValue> array = par[0].getArrayValue();
         ASTNode function = par[1].getExpressionNode();
 
-        return EvaluationValue.arrayValue(Lists.transform(array, input -> runLambda(expression, input, function)));
+        return EvaluationValue.arrayValue(Lists.transform(array, input -> Exceptions.supply(() -> runLambda(expression, input, function))));
     }
 
     @Override
-    public EvaluationValue cmd$inlineFunction(Expression expression, ASTNode node) throws ParseException, EvaluationException {
-        var value = CustomInlinerFunction.getNodeValue(node.getParameters().get(0));
-        if (value == null) return null;
-        return evaluate(expression, node.getToken(), value, EvaluationValue.expressionNodeValue(node.getParameters().get(1)));
+    public EvaluationValue inlineFunction(Expression expression, Token token, List<ASTNode> parameters) throws EvaluationException {
+        return null;
     }
 }