+++ src/main/java/me/melontini/commander/impl/expression/functions/math/ClampFunction.java
@@ -1,7 +1,7 @@
 package me.melontini.commander.impl.expression.functions.math;
 
+import com.ezylang.evalex.EvaluationContext;
 import com.ezylang.evalex.EvaluationException;
-import com.ezylang.evalex.Expression;
 import com.ezylang.evalex.data.EvaluationValue;
 import com.ezylang.evalex.functions.AbstractFunction;
 import com.ezylang.evalex.functions.FunctionParameter;
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 @FunctionParameter(name = "max")
 public class ClampFunction extends AbstractFunction {
     @Override
-    public EvaluationValue evaluate(Expression expression, Token functionToken, EvaluationValue... par) throws EvaluationException {
+    public EvaluationValue evaluate(EvaluationContext context, Token functionToken, EvaluationValue... par) throws EvaluationException {
         BigDecimal value = par[0].getNumberValue();
         BigDecimal min = par[1].getNumberValue();
         BigDecimal max = par[2].getNumberValue();