+++ src/main/java/me/melontini/commander/impl/expression/functions/math/RangedRandomFunction.java
@@ -1,30 +1,31 @@
 package me.melontini.commander.impl.expression.functions.math;
 
+import com.ezylang.evalex.EvaluationContext;
 import com.ezylang.evalex.EvaluationException;
 import com.ezylang.evalex.Expression;
 import com.ezylang.evalex.data.EvaluationValue;
 import com.ezylang.evalex.functions.AbstractFunction;
 import com.ezylang.evalex.functions.FunctionParameter;
 import com.ezylang.evalex.parser.ASTNode;
-import com.ezylang.evalex.parser.ParseException;
 import com.ezylang.evalex.parser.Token;
-import me.melontini.commander.impl.expression.functions.CustomInlinerFunction;
 import me.melontini.dark_matter.api.base.util.MathUtil;
+import org.jetbrains.annotations.Nullable;
 
 import java.math.BigDecimal;
+import java.util.List;
 
 @FunctionParameter(name = "min")
 @FunctionParameter(name = "max")
-public class RangedRandomFunction extends AbstractFunction implements CustomInlinerFunction {
+public class RangedRandomFunction extends AbstractFunction {
     @Override
-    public EvaluationValue evaluate(Expression expression, Token functionToken, EvaluationValue... par) throws EvaluationException {
+    public EvaluationValue evaluate(EvaluationContext context, Token functionToken, EvaluationValue... par) throws EvaluationException {
         BigDecimal min = par[0].getNumberValue();
         BigDecimal max = par[1].getNumberValue();
-        return expression.convertValue(min.compareTo(max) >= 0 ? min : BigDecimal.valueOf(MathUtil.threadRandom().nextDouble()).multiply(max.subtract(min)).add(min));
+        return context.expression().convertValue(min.compareTo(max) >= 0 ? min : BigDecimal.valueOf(MathUtil.threadRandom().nextDouble()).multiply(max.subtract(min)).add(min));
     }
 
     @Override
-    public EvaluationValue cmd$inlineFunction(Expression expression, ASTNode node) throws ParseException, EvaluationException {
+    public @Nullable EvaluationValue inlineFunction(Expression expression, Token token, List<ASTNode> parameters) throws EvaluationException {
         return null;
     }
 }