+++ src/main/java/me/melontini/commander/impl/expression/macro/DynamicMacro.java
@@ -3,12 +3,13 @@ package me.melontini.commander.impl.expression.macro;
 import me.melontini.commander.api.expression.BrigadierMacro;
 import net.minecraft.loot.context.LootContext;
 
-import java.util.function.Function;
+import java.util.Map;
+import java.util.function.BiFunction;
 
-public record DynamicMacro(String original, Function<LootContext, StringBuilder> start) implements BrigadierMacro {
+public record DynamicMacro(String original, BiFunction<LootContext, Map<String, Object>, StringBuilder> start) implements BrigadierMacro {
 
     @Override
-    public String build(LootContext context) {
-        return start.apply(context).toString();
+    public String build(LootContext context, Map<String, Object> params) {
+        return start.apply(context, params).toString();
     }
 }