+++ CHANGELOG.md
@@ -28,7 +28,7 @@ Dev Changes:
 There's a handful new experimental APIs, which allow you to extend the expression system.
 
 * Exposed `Object getValue()` and `Result convert(Object o)` in Expression.Result.
-* Exposed `Expression variable(String variable, Object value)` in Expression.
+* Exposed `Expression eval(LootContext context, Map<String, Object> parameters)` in Expression.
 * Exposed `ProxyMap` and `ObjectConverter`.
 * Exposed `CustomFields`. One of the more interesting APIs. Allows adding new 'virtual' fields to reflective expression objects.
   ```java